Removed code that should not exist anymore
