+++ ega-data-api-dataedge/src/main/java/eu/elixir/ega/ebi/dataedge/rest/FileController.java
@@ -35,9 +35,6 @@ import java.util.List;
 
 import static org.springframework.web.bind.annotation.RequestMethod.*;
 
-/**
- * @author asenf
- */
 @RestController
 @EnableDiscoveryClient
 @Slf4j
@@ -49,13 +46,6 @@ public class FileController {
 
     @Autowired
     private AuthenticationService authenticationService;
-
-    /**
-     * Writes a requested file, or part of file from the FileService to the supplied
-     * response stream.
-     *
-
-     */
 
     /**
      *
@@ -126,185 +116,6 @@ public class FileController {
                 destinationFormat,
                 request,
                 response);
-    }
-
-    /**
-     * Experimental - Return a BAM Header.
-     *
-     * @param fileId            ELIXIR id of the requested file.
-     * @param destinationFormat Requested destination format.
-     * @param destinationKey    Encryption key that the result file will be
-     *                          encrypted with.
-     * @return The BAM file header for the file.
-     */
-    @RequestMapping(value = "/{fileId}/header", method = GET)
-    public Object getFileHeader(@PathVariable String fileId,
-                                @RequestParam(value = "destinationFormat", required = false, defaultValue = "aes128") String destinationFormat,
-                                @RequestParam(value = "destinationKey", required = false, defaultValue = "") String destinationKey) {
-
-        return fileService.getFileHeader(fileId,
-                destinationFormat,
-                destinationKey,
-                null);  // This by default makes it BAM
-
-    }
-
-    /**
-     *
-     * @param response
-     */
-    @RequestMapping(value = "/byid/{type}", method = OPTIONS)
-    public void getById_(HttpServletResponse response) {
-        response.addHeader("Access-Control-Request-Method", "GET");
-    }
-
-    // {id} -- 'file', 'sample', 'run', ...
-    /**
-     * Writes a requested file (or part of file), selected by accession, from the
-     * FileService to the supplied response stream.
-     *
-     * @param type              Should be set to 'file'.
-     * @param accession         Local accession ID of the requested file.
-     * @param format            Requested file format. Either 'bam' or 'cram' (case
-     *                          insensitive).
-     * @param reference         FASTA reference name, required for selecting a
-     *                          region with start and end.
-     * @param start             Start coordinate when requesting a partial file.
-     * @param end               End coordinate when requesting a partial file.
-     * @param fields            Data fields to include in the output file.
-     * @param tags              Data tags to include in the output file.
-     * @param notags            Data tags to exclude from the output file.
-     * @param header            Unused.
-     * @param destinationFormat Requested destination format.
-     * @param destinationKey    Unused.
-     * @param request           Unused.
-     * @param response          Response stream for the returned data.
-     */
-    @RequestMapping(value = "/byid/{type}", method = GET)
-    @ResponseBody
-    public void getById(@PathVariable String type,
-                        @RequestParam(value = "accession") String accession,
-                        @RequestParam(value = "format", required = false, defaultValue = "bam") String format,
-                        @RequestParam(value = "chr", required = false, defaultValue = "") String reference,
-                        @RequestParam(value = "start", required = false, defaultValue = "0") long start,
-                        @RequestParam(value = "end", required = false, defaultValue = "0") long end,
-                        @RequestParam(value = "fields", required = false) List<String> fields,
-                        @RequestParam(value = "tags", required = false) List<String> tags,
-                        @RequestParam(value = "notags", required = false) List<String> notags,
-                        @RequestParam(value = "header", required = false, defaultValue = "true") Boolean header,
-                        @RequestParam(value = "destinationFormat", required = false, defaultValue = "plain") String destinationFormat,
-                        @RequestParam(value = "destinationKey", required = false, defaultValue = "") String destinationKey,
-                        HttpServletRequest request,
-                        HttpServletResponse response) {
-        Authentication auth = authenticationService.getAuthentication();
-        if (auth == null) {
-            throw new InvalidAuthenticationException(accession);
-        }
-        fileService.getById(type,
-                accession,
-                format,
-                reference,
-                start,
-                end,
-                fields,
-                tags,
-                notags,
-                header,
-                destinationFormat,
-                destinationKey,
-                request,
-                response);
-    }
-
-    /**
-     *
-     * @param response
-     */
-    @RequestMapping(value = "/variant/byid/{type}", method = OPTIONS)
-    public void getByVariantId_(HttpServletResponse response) {
-        response.addHeader("Access-Control-Request-Method", "GET");
-    }
-
-    /**
-     * Writes a requested file (or part of file), selected by accession, from the
-     * FileService to the supplied response stream.
-     *
-     * @param type              Should be set to 'file'.
-     * @param accession         Local accession ID of the requested file.
-     * @param format            Unused.
-     * @param reference         FASTA reference name, required for selecting a
-     *                          region with start and end.
-     * @param start             Start coordinate when requesting a partial file.
-     * @param end               End coordinate when requesting a partial file.
-     * @param fields            Data fields to include in the output file.
-     * @param tags              Data tags to include in the output file.
-     * @param notags            Data tags to exclude from the output file.
-     * @param header            Unused.
-     * @param destinationFormat Requested destination format.
-     * @param destinationKey    Unused.
-     * @param request           Unused.
-     * @param response          Response stream for the returned data.
-     */
-    @RequestMapping(value = "/variant/byid/{type}", method = GET)
-    @ResponseBody
-    public void getByVariantId(@PathVariable String type,
-                               @RequestParam(value = "accession") String accession,
-                               @RequestParam(value = "format", required = false, defaultValue = "vcf") String format,
-                               @RequestParam(value = "chr", required = false, defaultValue = "") String reference,
-                               @RequestParam(value = "start", required = false, defaultValue = "0") long start,
-                               @RequestParam(value = "end", required = false, defaultValue = "0") long end,
-                               @RequestParam(value = "fields", required = false) List<String> fields,
-                               @RequestParam(value = "tags", required = false) List<String> tags,
-                               @RequestParam(value = "notags", required = false) List<String> notags,
-                               @RequestParam(value = "header", required = false, defaultValue = "true") Boolean header,
-                               @RequestParam(value = "destinationFormat", required = false, defaultValue = "plain") String destinationFormat,
-                               @RequestParam(value = "destinationKey", required = false, defaultValue = "") String destinationKey,
-                               HttpServletRequest request,
-                               HttpServletResponse response) {
-        Authentication auth = authenticationService.getAuthentication();
-        if (auth == null) {
-            throw new InvalidAuthenticationException(accession);
-        }
-        fileService.getVCFById(type,
-                accession,
-                format,
-                reference,
-                start,
-                end,
-                fields,
-                tags,
-                notags,
-                header,
-                destinationFormat,
-                destinationKey,
-                request,
-                response);
-    }
-
-    /**
-     * Writes the content length of a selected file to the reponse parameter, and
-     * returns OK or UNAUTHORIZED wheather the file exists and can be accessible.
-     *
-     * @param fileId    should be "file".
-     * @param accession accession id of the requested file.
-     * @param request   Unused.
-     * @param response  reponse object which will be modified with the content
-     *                  length of the requested file head.
-     * @return httpStatus OK if the file info was accessible, and the reponse was
-     *         modified, otherwise UNAUTHORIZED.
-     */
-    @RequestMapping(value = "/byid/{type}", method = HEAD)
-    @ResponseBody
-    public ResponseEntity getHeadById(@PathVariable String type,
-                                      @RequestParam(value = "accession") String accession,
-                                      HttpServletRequest request,
-                                      HttpServletResponse response) {
-        Authentication auth = authenticationService.getAuthentication();
-        if (auth == null) {
-            throw new InvalidAuthenticationException(accession);
-        }
-
-        return fileService.getHeadById(type, accession, request, response);
     }
 
     /**
