+++ ega-data-api-dataedge/src/main/java/eu/elixir/ega/ebi/dataedge/service/FileService.java
@@ -15,16 +15,11 @@
  */
 package eu.elixir.ega.ebi.dataedge.service;
 
-import htsjdk.samtools.cram.ref.CRAMReferenceSource;
 import org.springframework.http.ResponseEntity;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.util.List;
 
-/**
- * @author asenf
- */
 public interface FileService {
 
     /**
@@ -65,93 +60,6 @@ public interface FileService {
                      String destinationFormat,
                      HttpServletRequest request,
                      HttpServletResponse response);
-
-    /**
-     * Returns the SAM file header for a file identified by fileId.
-     *
-     * @param fileId ELIXIR id of the requested file.
-     * @param destinationFormat Requested destination format.
-     * @param destinationKey Encryption key that the result file will be
-     *     encrypted with.
-     * @param x optional CRAM reference source to be used with the
-     *     SamReaderFactory.
-     * @return The SAM file header for the file.
-     */
-    Object getFileHeader(String fileId,
-                         String destinationFormat,
-                         String destinationKey,
-                         CRAMReferenceSource x);
-
-    /**
-     * Writes a requested file (or part of file), selected by accession, from the
-     * FileService to the supplied response stream.
-     *
-     * @param fileId            Should be set to 'file'.
-     * @param accession         Local accession ID of the requested file.
-     * @param format            Requested file format. Either 'bam' or 'cram' (case
-     *                          insensitive).
-     * @param reference         FASTA reference name, required for selecting a
-     *                          region with start and end.
-     * @param start             Start coordinate when requesting a partial file.
-     * @param end               End coordinate when requesting a partial file.
-     * @param fields            Data fields to include in the output file.
-     * @param tags              Data tags to include in the output file.
-     * @param notags            Data tags to exclude from the output file.
-     * @param header            Unused.
-     * @param destinationFormat Requested destination format.
-     * @param destinationKey    Unused.
-     * @param request           Unused.
-     * @param response          Response stream for the returned data.
-     */
-    void getById(String fileId,
-                 String accession,
-                 String format,
-                 String reference,
-                 long start,
-                 long end,
-                 List<String> fields,
-                 List<String> tags,
-                 List<String> notags,
-                 boolean header,
-                 String destinationFormat,
-                 String destinationKey,
-                 HttpServletRequest request,
-                 HttpServletResponse response);
-
-    /**
-     * Writes a requested file (or part of file), selected by accession, from the
-     * FileService to the supplied response stream.
-     *
-     * @param fileId            Should be set to 'file'.
-     * @param accession         Local accession ID of the requested file.
-     * @param format            Unused.
-     * @param reference         FASTA reference name, required for selecting a
-     *                          region with start and end.
-     * @param start             Start coordinate when requesting a partial file.
-     * @param end               End coordinate when requesting a partial file.
-     * @param fields            Data fields to include in the output file.
-     * @param tags              Data tags to include in the output file.
-     * @param notags            Data tags to exclude from the output file.
-     * @param header            Unused.
-     * @param destinationFormat Requested destination format.
-     * @param destinationKey    Unused.
-     * @param request           Unused.
-     * @param response          Response stream for the returned data.
-     */
-    void getVCFById(String fileId,
-                    String accession,
-                    String format,
-                    String reference,
-                    long start,
-                    long end,
-                    List<String> fields,
-                    List<String> tags,
-                    List<String> notags,
-                    boolean header,
-                    String destinationFormat,
-                    String destinationKey,
-                    HttpServletRequest request,
-                    HttpServletResponse response);
 
     /**
      * Writes the content length of a selected file to the reponse parameter, and
