+++ ega-data-api-htsget/src/main/java/eu/elixir/ega/ebi/htsget/config/OAuth2ResourceConfig.java
@@ -65,7 +65,6 @@ public class OAuth2ResourceConfig extends ResourceServerConfigurerAdapter {
                 .requestMatchers()
                 .antMatchers("/tickets/**")
                 .antMatchers("/htsget/**")
-                //.antMatchers("/stats/testme")
                 .and()
                 .authorizeRequests()
                 .antMatchers(HttpMethod.OPTIONS, "/**").permitAll()
