+++ ega-data-api-htsget/src/main/java/eu/elixir/ega/ebi/htsget/config/WebMvcConfig.java
@@ -1,20 +1,16 @@
 package eu.elixir.ega.ebi.htsget.config;
 
-import eu.elixir.ega.ebi.commons.config.InvalidAuthenticationException;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.MediaType;
 import org.springframework.web.servlet.config.annotation.ContentNegotiationConfigurer;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 @Configuration
 public class WebMvcConfig extends WebMvcConfigurerAdapter {
+
     @Override
     public void configureContentNegotiation(ContentNegotiationConfigurer configurer) {
         configurer.defaultContentType(new MediaType("application", "vnd.ga4gh.htsget.v1.0.0+json"));
     }
+
 }
