--- ega-data-api-htsget/src/main/java/eu/elixir/ega/ebi/htsget/dto/HtsgetUrl.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2017 ELIXIR EGA
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package eu.elixir.ega.ebi.htsget.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-/**
- * @author asenf
- */
-@NoArgsConstructor
-@AllArgsConstructor
-@Setter
-@Getter
-public class HtsgetUrl {
-
-    private String url;
-    private HtsgetHeader headers;
-
-}
