+++ ega-data-api-htsget/src/main/java/eu/elixir/ega/ebi/htsget/dto/HtsgetUrlV2.java
@@ -6,6 +6,7 @@ import java.net.URI;
 import java.util.HashMap;
 
 public class HtsgetUrlV2 {
+
     private URI url;
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private HashMap<String, String> headers;
