+++ ega-data-api-htsget/src/main/java/eu/elixir/ega/ebi/htsget/rest/TicketController.java
@@ -1,115 +1,150 @@
-/*
- * Copyright 2017 ELIXIR EGA
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package eu.elixir.ega.ebi.htsget.rest;
 
+import eu.elixir.ega.ebi.commons.config.HtsgetException;
+import eu.elixir.ega.ebi.commons.config.InvalidAuthenticationException;
+import eu.elixir.ega.ebi.commons.config.UnsupportedFormatException;
+import eu.elixir.ega.ebi.commons.exception.NotFoundException;
+import eu.elixir.ega.ebi.commons.exception.PermissionDeniedException;
+import eu.elixir.ega.ebi.htsget.dto.HtsgetErrorResponse;
+import eu.elixir.ega.ebi.htsget.dto.HtsgetResponseV2;
+import eu.elixir.ega.ebi.htsget.dto.HtsgetUrlV2;
 import eu.elixir.ega.ebi.htsget.service.TicketService;
+import org.apache.http.HttpHeaders;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.ServletRequestBindingException;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.net.URISyntaxException;
 import java.util.List;
+import java.util.Optional;
 
 import static org.springframework.web.bind.annotation.RequestMethod.GET;
-import static org.springframework.web.bind.annotation.RequestMethod.OPTIONS;
 
 @RestController
-@EnableDiscoveryClient
-@RequestMapping("/tickets")
+//@EnableDiscoveryClient
+@RequestMapping({"/htsget","/tickets"})
+@CrossOrigin(methods = GET, maxAge = 30 * 24 * 60 * 60)
 public class TicketController {
 
     @Autowired
-    private TicketService ticketService;
+    private TicketService service;
 
-    @RequestMapping(value = "/files/{fileId}", method = OPTIONS)
-    public void getTicket(HttpServletResponse response) {
-        response.addHeader("Access-Control-Request-Method", "GET");
+    @RequestMapping(value = "/version", method = GET)
+    @ResponseBody
+    public String getVersion() {
+        return "v1.0.0";
     }
 
-    @RequestMapping(value = "/files/{fileId}", method = GET)
-    public Object getTicket(@PathVariable String fileId,
-                            @RequestParam(name = "format", required = false, defaultValue = "BAM") String format,
-                            @RequestParam(name = "referenceIndex", required = false, defaultValue = "-1") Integer referenceIndex,
-                            @RequestParam(name = "referenceName", required = false, defaultValue = "") String referenceName,
-                            @RequestParam(name = "referenceMD5", required = false, defaultValue = "") String referenceMD5,
-                            @RequestParam(name = "start", required = false, defaultValue = "") String start,
-                            @RequestParam(name = "end", required = false, defaultValue = "") String end,
-                            @RequestParam(name = "fields", required = false) List<String> fields,
-                            @RequestParam(name = "tags", required = false) List<String> tags,
-                            @RequestParam(name = "notags", required = false) List<String> notags,
-                            HttpServletRequest request,
-                            HttpServletResponse response) {
-
-        return ticketService.getTicket(fileId,
-                format,
-                referenceIndex,
-                referenceName,
-                referenceMD5,
-                start,
-                end,
-                fields,
-                tags,
-                notags,
-                request,
-                response);
+    @RequestMapping(value = "/reads/{id}", method = GET, produces = "application/vnd.ga4gh.htsget.v1.0.0+json")
+    public HtsgetResponseV2 getRead(@PathVariable String id,
+                                    @RequestParam(defaultValue = "BAM") String format,
+                                    @RequestParam(name = "class") Optional<String> requestClass,
+                                    @RequestParam Optional<String> referenceName,
+                                    @RequestParam Optional<Long> start,
+                                    @RequestParam Optional<Long> end,
+                                    @RequestParam Optional<List<TicketService.Field>> fields,
+                                    @RequestParam Optional<List<String>> tags,
+                                    @RequestParam Optional<List<String>> notags,
+                                    @RequestHeader String authorization
+    ) throws IOException, URISyntaxException {
+        HtsgetResponseV2 response = service.getRead(id, format, requestClass, referenceName, start, end, fields, tags, notags);
+        for (HtsgetUrlV2 url : response.getUrls()) {
+            url.setHeader(HttpHeaders.AUTHORIZATION, authorization);
+        }
+        return response;
     }
 
-    @RequestMapping(value = "/variants/{fileId}", method = OPTIONS)
-    public void getVariant(HttpServletResponse response) {
-        response.addHeader("Access-Control-Request-Method", "GET");
+    @RequestMapping(value = "/variants/{id}", method = GET)
+    public HtsgetResponseV2 getVariant(@PathVariable String id,
+                                       @RequestParam(defaultValue = "VCF") String format,
+                                       @RequestParam(name = "class") Optional<String> requestClass,
+                                       @RequestParam Optional<String> referenceName,
+                                       @RequestParam Optional<Long> start,
+                                       @RequestParam Optional<Long> end,
+                                       @RequestParam Optional<List<TicketService.Field>> fields,
+                                       @RequestParam Optional<List<String>> tags,
+                                       @RequestParam Optional<List<String>> notags,
+                                       @RequestHeader String authorization) throws IOException, URISyntaxException {
+        HtsgetResponseV2 response = service.getVariant(id, format, requestClass, referenceName, start, end, fields, tags, notags);
+        for (HtsgetUrlV2 url : response.getUrls()) {
+            url.setHeader(HttpHeaders.AUTHORIZATION, authorization);
+        }
+        return response;
     }
 
-    @RequestMapping(value = "/variants/{fileId}", method = GET)
-    public Object getVariantTicket(@PathVariable String fileId,
-                                   @RequestParam(name = "format", required = false, defaultValue = "VCF") String format,
-                                   @RequestParam(name = "referenceIndex", required = false, defaultValue = "-1") Integer referenceIndex,
-                                   @RequestParam(name = "referenceName", required = false, defaultValue = "") String referenceName,
-                                   @RequestParam(name = "referenceMD5", required = false, defaultValue = "") String referenceMD5,
-                                   @RequestParam(name = "start", required = false, defaultValue = "") String start,
-                                   @RequestParam(name = "end", required = false, defaultValue = "") String end,
-                                   @RequestParam(name = "fields", required = false) List<String> fields,
-                                   @RequestParam(name = "tags", required = false) List<String> tags,
-                                   @RequestParam(name = "notags", required = false) List<String> notags,
-                                   HttpServletRequest request,
-                                   HttpServletResponse response) {
+    @RequestMapping(value = "/files/{id}", method = GET)
+    public HtsgetResponseV2 getFile(@PathVariable String id,
+                                    @RequestParam(defaultValue = "BAM") String format,
+                                    @RequestParam(name = "class") Optional<String> requestClass,
+                                    @RequestParam Optional<String> referenceName,
+                                    @RequestParam Optional<Long> start,
+                                    @RequestParam Optional<Long> end,
+                                    @RequestParam Optional<List<TicketService.Field>> fields,
+                                    @RequestParam Optional<List<String>> tags,
+                                    @RequestParam Optional<List<String>> notags,
+                                    @RequestHeader String authorization) throws IOException, URISyntaxException {
+        HtsgetResponseV2 response;
+        if (format.equalsIgnoreCase("BAM") || format.equalsIgnoreCase("CRAM")) {
+            response = service.getRead(id, format, requestClass, referenceName, start, end, fields, tags, notags);
+        } else if (format.equalsIgnoreCase("VCF") || format.equalsIgnoreCase("BCF")) {
+            response = service.getVariant(id, format, requestClass, referenceName, start, end, fields, tags, notags);
+        } else {
+            throw new UnsupportedFormatException(format);
+        }
 
-        return ticketService.getVariantTicket(fileId,
-                format,
-                referenceIndex,
-                referenceName,
-                referenceMD5,
-                start,
-                end,
-                fields,
-                tags,
-                notags,
-                request,
-                response);
+        for (HtsgetUrlV2 url : response.getUrls()) {
+            url.setHeader(HttpHeaders.AUTHORIZATION, authorization);
+        }
+        return response;
     }
 
-    @RequestMapping(value = "/**", method = RequestMethod.OPTIONS)
-    public ResponseEntity handle() {
-        return new ResponseEntity(HttpStatus.OK);
+    @ExceptionHandler
+    public ResponseEntity<HtsgetErrorResponse> handleHtsgetException(HtsgetException exception) {
+        return ResponseEntity
+                .status(exception.getStatusCode())
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(new HtsgetErrorResponse(exception.getHtsgetErrorCode(), exception.getMessage()));
     }
 
+    @ExceptionHandler
+    public ResponseEntity<HtsgetErrorResponse> handlePermissionDeniedException(PermissionDeniedException exception) {
+        return ResponseEntity
+                .status(403)
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(new HtsgetErrorResponse("PermissionDenied", exception.getMessage()));
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<HtsgetErrorResponse> handleNotFoundException(NotFoundException exception) {
+        return ResponseEntity
+                .status(404)
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(new HtsgetErrorResponse("NotFound", exception.getMessage()));
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<HtsgetErrorResponse> handleInvalidAuthenticationException(InvalidAuthenticationException exception) {
+        return ResponseEntity
+                .status(401)
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(new HtsgetErrorResponse("InvalidAuthentication", exception.getMessage()));
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<HtsgetErrorResponse> handleServletRequestBindingException(ServletRequestBindingException exception) {
+        return ResponseEntity
+                .status(401)
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(new HtsgetErrorResponse("InvalidAuthentication", "Request missing Authorization header"));
+    }
 }
