+++ ega-data-api-htsget/src/main/java/eu/elixir/ega/ebi/htsget/service/TicketService.java
@@ -1,51 +1,50 @@
-/*
- * Copyright 2017 ELIXIR EGA
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package eu.elixir.ega.ebi.htsget.service;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.util.List;
+import eu.elixir.ega.ebi.commons.config.HtsgetException;
+import eu.elixir.ega.ebi.commons.exception.NotFoundException;
+import eu.elixir.ega.ebi.commons.exception.PermissionDeniedException;
+import eu.elixir.ega.ebi.htsget.dto.HtsgetResponseV2;
 
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.util.List;
+import java.util.Optional;
 
 public interface TicketService {
 
-    Object getTicket(String fileId,
-                     String format,
-                     int referenceIndex,
-                     String referenceName,
-                     String referenceMD5,
-                     String start,
-                     String end,
-                     List<String> fields,
-                     List<String> tags,
-                     List<String> notags,
-                     HttpServletRequest request,
-                     HttpServletResponse response);
+    HtsgetResponseV2 getRead(String id,
+                             String format,
+                             Optional<String> requestClass,
+                             Optional<String> referenceName,
+                             Optional<Long> start,
+                             Optional<Long> end,
+                             Optional<List<Field>> fields,
+                             Optional<List<String>> tags,
+                             Optional<List<String>> notags)
+            throws HtsgetException, NotFoundException, PermissionDeniedException, IOException, URISyntaxException;
 
-    Object getVariantTicket(String fileId,
-                            String format,
-                            int referenceIndex,
-                            String referenceName,
-                            String referenceMD5,
-                            String start,
-                            String end,
-                            List<String> fields,
-                            List<String> tags,
-                            List<String> notags,
-                            HttpServletRequest request,
-                            HttpServletResponse response);
-    
+    HtsgetResponseV2 getVariant(String id,
+                                String format,
+                                Optional<String> requestClass,
+                                Optional<String> referenceName,
+                                Optional<Long> start,
+                                Optional<Long> end,
+                                Optional<List<Field>> fields,
+                                Optional<List<String>> tags,
+                                Optional<List<String>> notags)
+            throws HtsgetException, NotFoundException, PermissionDeniedException, IOException, URISyntaxException;
+
+    enum Field {
+        QNAME,
+        FLAG,
+        RNAME,
+        POS,
+        MAPQ,
+        CIGAR,
+        RNEXT,
+        PNEXT,
+        TLEN,
+        SEQ,
+        QUAL
+    }
 }
